+++ app/src/main/java/org/breezyweather/main/adapters/HomePollenAdapter.kt
@@ -24,6 +24,7 @@ import org.breezyweather.R
 import breezyweather.domain.location.model.Location
 import breezyweather.domain.weather.model.Daily
 import org.breezyweather.common.extensions.getFormattedDate
+import org.breezyweather.common.source.PollenIndexSource
 import org.breezyweather.common.ui.composables.PollenGrid
 import org.breezyweather.databinding.ItemPollenDailyBinding
 import org.breezyweather.domain.weather.model.isIndexValid
@@ -31,7 +32,8 @@ import org.breezyweather.main.utils.MainThemeColorProvider
 import org.breezyweather.theme.compose.BreezyWeatherTheme
 
 open class HomePollenAdapter(
-    private val location: Location
+    private val location: Location,
+    private val pollenIndexSource: PollenIndexSource?
 ) : RecyclerView.Adapter<HomePollenViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): HomePollenViewHolder {
@@ -43,10 +45,16 @@ open class HomePollenAdapter(
     }
 
     override fun onBindViewHolder(holder: HomePollenViewHolder, position: Int) {
-        holder.onBindView(location, location.weather!!.dailyForecastStartingToday[position])
+        holder.onBindView(
+            location,
+            location.weather!!.dailyForecastStartingToday[position],
+            pollenIndexSource
+        )
     }
 
-    override fun getItemCount() = location.weather?.dailyForecastStartingToday?.filter { it.pollen?.isIndexValid == true }?.size ?: 0
+    override fun getItemCount() = location.weather?.dailyForecastStartingToday?.filter {
+        it.pollen?.isIndexValid == true
+    }?.size ?: 0
 }
 
 class HomePollenViewHolder internal constructor(
@@ -55,7 +63,7 @@ class HomePollenViewHolder internal constructor(
     binding.root
 ) {
     @SuppressLint("SetTextI18n", "RestrictedApi")
-    fun onBindView(location: Location, daily: Daily) {
+    fun onBindView(location: Location, daily: Daily, pollenIndexSource: PollenIndexSource?) {
         val context = itemView.context
 
         binding.title.text = daily.date.getFormattedDate(location.timeZone, context.getString(R.string.date_format_widget_long))
@@ -64,7 +72,10 @@ class HomePollenViewHolder internal constructor(
         daily.pollen?.let {
             binding.composeView.setContent {
                 BreezyWeatherTheme(lightTheme = MainThemeColorProvider.isLightTheme(context, location)) {
-                    PollenGrid(pollen = it)
+                    PollenGrid(
+                        pollen = it,
+                        pollenIndexSource = pollenIndexSource
+                    )
                 }
             }
         }